move areas components into their own folders
