+++ src/components/organisms/RequestArea/components/RequestHeaders.jsx
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
-import useAppContext from '../../../state/useAppContext';
-import DynamicTable from '../../templates/DynamicTable';
+import useAppContext from '../../../../state/useAppContext';
+import NoDataMessage from '../../../molecules/NoDataMessage';
+import DynamicTable from '../../../templates/DynamicTable';
 
 function HeadersTable() {
   const { setRequestElement } = useAppContext();
@@ -13,6 +14,7 @@ function HeadersTable() {
         tableKey="request-headers"
         buttonText="Add a header"
         onTableChange={(headers) => setRequestElement('headers', headers)}
+        noDataMessage={<NoDataMessage icon="align-top-simple" />}
       />
     </StyledParams>
   );
